% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpChecks.R
\name{addIfAny}
\alias{addIfAny}
\title{Helper function to conditionally add elements to a list}
\usage{
addIfAny(validation, name, value)
}
\arguments{
\item{validation}{A list to which elements may be added.}

\item{name}{A character string representing the name of the element to add.}

\item{value}{The value to add to the list if it is not NULL or empty}
}
\value{
The updated list with the new element added if applicable.
}
\description{
Helper function to conditionally add elements to a list
}
\keyword{internal}
