% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addParentalChain.R
\name{addMaternalLineFlag}
\alias{addMaternalLineFlag}
\title{Add a maternal-line descendant flag to a pedigree}
\usage{
addMaternalLineFlag(ped, anchor_id, flag_col, chain_col = "momID_chain")
}
\arguments{
\item{ped}{A pedigree data frame containing a maternal chain list-column.}

\item{anchor_id}{ID of the anchor individual to search for within each
person's maternal chain.}

\item{flag_col}{Character string giving the name of the logical output column
to add to `ped`.}

\item{chain_col}{Character string giving the name of the list-column
containing ordered maternal ancestor chains.}
}
\value{
A data frame with `flag_col` added. The flag is `TRUE` when
  `anchor_id` appears in the individual's maternal chain and `FALSE`
  otherwise.
}
\description{
Adds a logical flag indicating whether a specified anchor individual appears
anywhere in each person's ordered maternal ancestor chain.
}
\details{
This is a convenience wrapper around [addParentalLineFlag()] with
`component = "momID"`.
}
