% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addParentalChain.R
\name{addPaternalLineFlag}
\alias{addPaternalLineFlag}
\title{Add a paternal-line descendant flag to a pedigree}
\usage{
addPaternalLineFlag(ped, anchor_id, flag_col, chain_col = "dadID_chain")
}
\arguments{
\item{ped}{A pedigree data frame containing a paternal chain list-column.}

\item{anchor_id}{ID of the anchor individual to search for within each
person's paternal chain.}

\item{flag_col}{Character string giving the name of the logical output column
to add to `ped`.}

\item{chain_col}{Character string giving the name of the list-column
containing ordered paternal ancestor chains.}
}
\value{
A data frame with `flag_col` added. The flag is `TRUE` when
  `anchor_id` appears in the individual's paternal chain and `FALSE`
  otherwise.
}
\description{
Adds a logical flag indicating whether a specified anchor individual appears
anywhere in each person's ordered paternal ancestor chain.
}
\details{
This is a convenience wrapper around [addParentalLineFlag()] with
`component = "dadID"`.
}
