% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkParents.R
\name{addPhantoms}
\alias{addPhantoms}
\title{Add Phantom Parents}
\usage{
addPhantoms(ped, verbose, pid, validation_results)
}
\arguments{
\item{ped}{A dataframe representing the pedigree data with columns 'ID', 'dadID', and 'momID'.}

\item{verbose}{A logical flag indicating whether to print progress and validation messages to the console.}
}
\description{
This function adds phantom parents to a pedigree.
}
