% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkParents.R
\name{addRowlessParents}
\alias{addRowlessParents}
\title{Add Phantom Parents}
\usage{
addRowlessParents(ped, verbose, validation_results)
}
\arguments{
\item{ped}{A dataframe representing the pedigree data with columns 'ID', 'dadID', and 'momID'.}

\item{verbose}{A logical flag indicating whether to print progress and validation messages to the console.}

\item{validation_results}{validation results}
}
\description{
This function adds parents w
}
