% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmxPedigrees.R
\name{alignPhenToOrdinal}
\alias{alignPhenToOrdinal}
\title{Align Phenotype Vector to Ordered Factor Data Frame}
\usage{
alignPhenToOrdinal(ped, phenotype, keep_ids, levels, personID = "ID")
}
\arguments{
\item{ped}{A data frame representing the pedigree.}

\item{phenotype}{A character string specifying the phenotype column name.}

\item{keep_ids}{A vector of IDs for which the phenotype values should be extracted.}

\item{levels}{A vector of the ordered levels for the factor (e.g., \code{c(0, 1)} for binary).}

\item{personID}{A character string specifying the ID column name. Default is "ID".}
}
\value{
A single-row data frame of ordered factors with columns named after \code{keep_ids}.
}
\description{
This function takes a pedigree data frame, a specified phenotype column, and a vector of IDs
to keep, and returns a single-row data frame of ordered factors suitable for OpenMx threshold
models (binary or ordinal).
}
