% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famSizeCal.R
\name{allGens}
\alias{allGens}
\title{allGens
A function to return a vector for number of individuals in each generation. This is a supporting function for function \code{SimPed}}
\usage{
allGens(kpc, Ngen, marR)
}
\arguments{
\item{kpc}{kids per couple}

\item{Ngen}{number of generations}

\item{marR}{Mating rate.}
}
\value{
Returns a vector including the number of individuals in every generation.
}
\description{
allGens
A function to return a vector for number of individuals in each generation. This is a supporting function for function \code{SimPed}
}
