% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpGeneric.R
\name{as_numeric_matrix}
\alias{as_numeric_matrix}
\title{Coerce to a Numeric Matrix}
\usage{
as_numeric_matrix(x)
}
\arguments{
\item{x}{An object coercible to a matrix.}
}
\value{
\code{x} as a matrix with numeric storage mode.
}
\description{
Coerce to a Numeric Matrix
}
\keyword{internal}
