% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpSimulatePedigree_beta.R
\name{assignCoupleIDs_beta}
\alias{assignCoupleIDs_beta}
\title{Assign Couple IDs}
\usage{
assignCoupleIDs_beta(df_Ngen)
}
\arguments{
\item{df_Ngen}{The dataframe for the current generation, including columns for individual IDs and spouse IDs.}
}
\value{
The input dataframe augmented with a 'coupleId' column, where each mated pair has a unique identifier.
}
\description{
This subfunction assigns a unique couple ID to each mated pair in the generation.
Unmated individuals are assigned NA for their couple ID.
}
\keyword{internal}
