% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGmisc-package.R
\name{BGmisc_package}
\alias{BGmisc_package}
\alias{BGmisc-package}
\title{Behavior Genetic Miscellaneous functions in R}
\description{
This collection contains functions for behavior genetic
  modeling. These functions include model identification, calculating
  relatedness, and various others.
}
\author{
S. Mason Garrison, Michael D. Hunter, S. Alexandra Burt, and Jonathan D. Trattner
}
