% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{calculateH}
\alias{calculateH}
\title{Falconer's Formula
Use Falconer's formula to solve for H using the observed correlations for two groups of any two relatednesses.}
\usage{
calculateH(r1, r2, obsR1, obsR2)
}
\arguments{
\item{r1}{Relatedness coefficient of the first group.}

\item{r2}{Relatedness coefficient of the second group.}

\item{obsR1}{Observed correlation between members of the first group.}

\item{obsR2}{Observed correlation between members of the second group.}
}
\value{
Heritability estimates (`heritability_estimates`).
}
\description{
Falconer's Formula
Use Falconer's formula to solve for H using the observed correlations for two groups of any two relatednesses.
}
