% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIDs.R
\name{checkIDuniqueness}
\alias{checkIDuniqueness}
\title{Check for duplicated individual IDs}
\usage{
checkIDuniqueness(ped, verbose = FALSE)
}
\arguments{
\item{ped}{A pedigree object}

\item{verbose}{A logical indicating whether to print progress messages}
}
\value{
A list containing the results of the check
}
\description{
This function checks for duplicated individual IDs in a pedigree.
}
