% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIDs.R
\name{checkWithinRowDuplicates}
\alias{checkWithinRowDuplicates}
\title{Check for within-row duplicates (self-parents, same mom/dad)}
\usage{
checkWithinRowDuplicates(ped, verbose = FALSE)
}
\description{
Check for within-row duplicates (self-parents, same mom/dad)
}
