% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLinks.R
\name{com2links.legacy}
\alias{com2links.legacy}
\title{com2links}
\usage{
com2links.legacy(
  rel_pairs_file = "dataRelatedPairs.csv",
  ad_ped_matrix = NULL,
  mit_ped_matrix = mt_ped_matrix,
  mt_ped_matrix = NULL,
  cn_ped_matrix = NULL,
  update_rate = 500,
  verbose = FALSE,
  outcome_name = "data",
  ...
)
}
\description{
This legacy function converts pedigree matrices into a related pairs file.
}
