% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLinks.R
\name{com2links}
\alias{com2links}
\title{Take a sparse component and turn it into kinship links}
\usage{
com2links(
  rel_pairs_file = "dataRelatedPairs.csv",
  ad_ped_matrix = NULL,
  mit_ped_matrix = mt_ped_matrix,
  mt_ped_matrix = NULL,
  cn_ped_matrix = NULL,
  write_buffer_size = 1000,
  update_rate = 1000,
  gc = TRUE,
  writetodisk = TRUE,
  verbose = FALSE,
  legacy = FALSE,
  outcome_name = "data",
  ...
)
}
\arguments{
\item{rel_pairs_file}{File to write related pairs to}

\item{ad_ped_matrix}{Matrix of additive genetic relatedness coefficients}

\item{mit_ped_matrix}{Matrix of mitochondrial relatedness coefficients}

\item{mt_ped_matrix}{(alternative) Matrix of mitochondrial relatedness coefficients}

\item{cn_ped_matrix}{Matrix of common nuclear relatedness coefficients}

\item{write_buffer_size}{Number of related pairs to write to disk at a time}

\item{update_rate}{numeric. How often to print progress messages}

\item{gc}{logical. If TRUE, do frequent garbage collection via \code{\link{gc}} to save memory}

\item{writetodisk}{logical. If TRUE, write the related pairs to disk}

\item{verbose}{logical. If TRUE, print progress messages}

\item{legacy}{logical. If TRUE, use the legacy version of the function}

\item{outcome_name}{Character string. The name of the outcome}

\item{...}{Additional arguments to be passed to \code{\link{com2links}}}
}
\value{
A data frame of related pairs
}
\description{
Take a sparse component and turn it into kinship links
}
