% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGedcom.R
\name{combine_columns}
\alias{combine_columns}
\title{Combine Columns}
\usage{
combine_columns(col1, col2)
}
\arguments{
\item{col1}{The first column to combine.}

\item{col2}{The second column to combine.}
}
\value{
A list with the combined column and a flag indicating if the second column should be retained.
}
\description{
This function combines two columns, handling conflicts and merging non-conflicting data.
}
\keyword{internal}
