% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{comp2vech}
\alias{comp2vech}
\title{Turn a variance component relatedness matrix into its half-vectorization}
\usage{
comp2vech(x, include.zeros = FALSE)
}
\arguments{
\item{x}{relatedness component matrix}

\item{include.zeros}{logical. Whether to include all-zero rows.}
}
\description{
Turn a variance component relatedness matrix into its half-vectorization
}
\details{
This is a wrapper around the \code{vech} function for producing the
half-vectorization of a matrix.  The extension here is to allow for
blockwise matrices.
}
\examples{
comp2vech(list(matrix(c(1, .5, .5, 1), 2, 2), matrix(1, 2, 2)))

}
