% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPedigree.R
\name{compute_transpose}
\alias{compute_transpose}
\title{Compute the transpose multiplication for the relatedness matrix}
\usage{
compute_transpose(r2, transpose_method = "tcrossprod", verbose = FALSE)
}
\arguments{
\item{r2}{a relatedness matrix}

\item{transpose_method}{character. The method to use for computing the transpose.  Options are "tcrossprod", "crossprod", or "star"}

\item{verbose}{logical.  If TRUE, print progress through stages of algorithm}
}
\description{
Compute the transpose multiplication for the relatedness matrix
}
\details{
The algorithms and methodologies used in this function are further discussed and exemplified in the vignette titled "examplePedigreeFunctions". For more advanced scenarios and detailed explanations, consult this vignette.
}
