% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGedcomlegacy.R
\name{countPatternRows.legacy}
\alias{countPatternRows.legacy}
\title{Check for Pattern Rows}
\usage{
countPatternRows.legacy(file)
}
\arguments{
\item{file}{A data frame containing the GEDCOM file.}
}
\value{
A list with the number of rows containing each pattern.
}
\description{
This function counts the number of rows containing specific patterns.
}
\keyword{internal}
