% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWikifamilytree.R
\name{deduplicatePairs}
\alias{deduplicatePairs}
\title{Deduplicate pairs of IDs in a data frame}
\usage{
deduplicatePairs(df)
}
\arguments{
\item{df}{A data frame with columns from_id and to_id}
}
\value{
A data frame with unique pairs of IDs
}
\description{
Deduplicate pairs of IDs in a data frame
}
