% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpPedigree.R
\name{determineSex}
\alias{determineSex}
\title{Determine Sex of Offspring}
\usage{
determineSex(idGen, sexR)
}
\arguments{
\item{idGen}{Vector of IDs for the generation.}

\item{sexR}{Numeric value indicating the sex ratio (proportion of males).}
}
\value{
Vector of sexes ("M" for male, "F" for female) for the offspring.
}
\description{
This function assigns sexes to the offspring in a generation based on the specified sex ratio.
}
