% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildComponent.R
\name{.assignParentValue}
\alias{.assignParentValue}
\title{Assign parent values based on component type}
\usage{
.assignParentValue(component)
}
\arguments{
\item{component}{character.  Which component of the pedigree to return.  See Details.}
}
\description{
Assign parent values based on component type
}
