% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGedcomlegacy.R
\name{.collapseNames.legacy}
\alias{.collapseNames.legacy}
\title{collapse Names}
\usage{
.collapseNames.legacy(verbose, df_temp)
}
\arguments{
\item{verbose}{A logical value indicating whether to print messages.}

\item{df_temp}{A data frame containing the columns to be combined.}
}
\description{
This function combines the `name_given` and `name_given_pieces` columns in a data frame.
}
\keyword{internal}
