% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGedcomlegacy.R
\name{.extract_info.legacy}
\alias{.extract_info.legacy}
\title{Extract Information from Line}
\usage{
.extract_info.legacy(line, type)
}
\arguments{
\item{line}{A character string representing a line from a GEDCOM file.}

\item{type}{A character string representing the type of information to extract.}
}
\value{
A character string with the extracted information.
}
\description{
This function extracts information from a line based on a specified type.
}
\keyword{internal}
