% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCIs.R
\name{.fisherz}
\alias{.fisherz}
\title{Fisher's r to z transformation and back}
\usage{
.fisherz(rho)
}
\arguments{
\item{rho}{A numeric vector of correlation coefficients.}

\item{z}{A numeric vector of Fisher's z-scores.}
}
\value{
A numeric vector of transformed values.
}
\description{
These functions convert correlation coefficients (r) to Fisher's z-scores and back.
}
\details{
Credit to the psych package for the Fisher's r to z transformation.
}
\keyword{internal}
