% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCIs.R
\name{.fisherz2r}
\alias{.fisherz2r}
\title{Fisher's r to z transformation and back}
\usage{
.fisherz2r(z)
}
\arguments{
\item{z}{A numeric vector of Fisher's z-scores.}
}
\value{
A numeric vector of transformed values.
}
\description{
Fisher's r to z transformation and back
}
\keyword{internal}
