% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSex.R
\name{.getModalValue}
\alias{.getModalValue}
\title{Get the Modal Value of a Vector}
\usage{
.getModalValue(x)
}
\arguments{
\item{x}{A vector of values.}
}
\value{
The modal value of the vector. If the vector is empty or contains only NA values, returns NA.
}
\description{
This function calculates the modal value of a vector, which is the most frequently occurring value.
If the vector is empty or contains only NA values, it returns NA.
}
\keyword{internal}
