% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildComponent.R
\name{.loadOrComputeInverseDiagonal}
\alias{.loadOrComputeInverseDiagonal}
\title{Load or compute the inverse diagonal matrix}
\usage{
.loadOrComputeInverseDiagonal(
  r,
  isChild,
  checkpoint_files,
  config,
  compress = TRUE
)
}
\arguments{
\item{r}{The relatedness matrix.}

\item{config}{A list containing configuration parameters such as `resume`, `verbose`, and `saveable`.}

\item{compress}{a logical specifying whether saving to a named file is to use "gzip" compression, or one of "gzip", "bzip2", "xz" or "zstd" to indicate the type of compression to be used. Ignored if file is a connection.}
}
\value{
The computed inverse diagonal matrix.
}
\description{
Load or compute the inverse diagonal matrix
}
\keyword{internal}
