% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildComponent.R
\name{.loadOrComputeIsChild}
\alias{.loadOrComputeIsChild}
\title{Load or compute the isChild matrix}
\usage{
.loadOrComputeIsChild(ped, checkpoint_files, config)
}
\arguments{
\item{ped}{a pedigree dataset.  Needs ID, momID, and dadID columns}

\item{checkpoint_files}{A list of checkpoint file paths.

 @keywords internal}

\item{config}{A list containing configuration parameters such as `resume`, `verbose`, and `saveable`.}
}
\description{
Load or compute the isChild matrix
}
