% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildComponent.R
\name{.loadOrComputeParList}
\alias{.loadOrComputeParList}
\title{parent-child adjacency data}
\usage{
.loadOrComputeParList(
  checkpoint_files,
  config,
  ped = NULL,
  parList = NULL,
  lens = NULL,
  compress = TRUE
)
}
\arguments{
\item{checkpoint_files}{A list of checkpoint file paths.}

\item{config}{A list containing configuration parameters such as `resume`, `verbose`, and `saveable`.}

\item{ped}{a pedigree dataset.  Needs ID, momID, and dadID columns}

\item{parList}{A list of parent-child adjacency data.}

\item{lens}{A vector of lengths for each parent-child relationship.}

\item{compress}{a logical specifying whether saving to a named file is to use "gzip" compression, or one of "gzip", "bzip2", "xz" or "zstd" to indicate the type of compression to be used. Ignored if file is a connection.}
}
\value{
A list containing the parent-child adjacency data either
        loaded from a checkpoint or initialized.
}
\description{
parent-child adjacency data
}
\keyword{internal}
