% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGedcomlegacy.R
\name{.mapFAMS2parents.legacy}
\alias{.mapFAMS2parents.legacy}
\title{Create a mapping of family IDs to parent IDs}
\usage{
.mapFAMS2parents.legacy(df_temp)
}
\arguments{
\item{df_temp}{A data frame containing information about individuals.}
}
\value{
A list mapping family IDs to parent IDs.
}
\description{
This function creates a mapping from family IDs to the IDs of the parents.
}
\keyword{internal}
