% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmxPedigrees.R
\name{.recycleFlag}
\alias{.recycleFlag}
\title{Recycle a logical flag to a required length (internal)}
\usage{
.recycleFlag(x, n, arg_name)
}
\arguments{
\item{x}{A logical vector of length 1 or \code{n}.}

\item{n}{Required length.}

\item{arg_name}{Name of the calling argument, used in error messages.}
}
\value{
A logical vector of length \code{n}.
}
\description{
Recycle a logical flag to a required length (internal)
}
\keyword{internal}
