% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpGeneric.R
\name{.require_openmx}
\alias{.require_openmx}
\title{Check for OpenMx Package}
\usage{
.require_openmx(fn = NULL)
}
\arguments{
\item{fn}{An optional character string representing the function name that requires OpenMx. This is included in the error message for clarity if provided.}
}
\value{
Returns `TRUE` invisibly if OpenMx is available. If OpenMx is not installed, it throws an error.
}
\description{
This function checks if the OpenMx package is installed and available for use. If the package is not installed, it throws an error with a message indicating that OpenMx is required and provides instructions for installation.
}
\keyword{internal}
