% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpChecks.R
\name{dropIdenticalDuplicateIDs}
\alias{dropIdenticalDuplicateIDs}
\title{Drop Identical Duplicate IDs from Pedigree Data Frame}
\usage{
dropIdenticalDuplicateIDs(ped, ids, changes = NULL)
}
\arguments{
\item{ped}{A data frame representing the pedigree.}

\item{ids}{A vector of IDs to check for duplicates in the pedigree.}
}
\description{
#' This function identifies and removes duplicate entries in a pedigree data frame
based on a list of specified IDs. If multiple rows share the same ID and are
identical, only one instance is retained. The function returns the modified pedigree
data frame along with a log of changes made.
}
