% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efunc.R
\name{efunc}
\alias{efunc}
\title{Error Function}
\usage{
efunc(error)
}
\arguments{
\item{error}{error output}
}
\value{
Replaces error message (\code{error}) with NA
}
\description{
Error Function
}
\keyword{internal}
