% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evenInsert.R
\name{evenInsert}
\alias{evenInsert}
\title{evenInsert
A function to insert m elements in a length n vector evenly. This is a supporting function for function \code{SimPed}}
\usage{
evenInsert(m, n)
}
\arguments{
\item{m}{a vector}

\item{n}{a vector}
}
\value{
Returns the inserted vector.
}
\description{
evenInsert
A function to insert m elements in a length n vector evenly. This is a supporting function for function \code{SimPed}
}
