% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGedcom.R
\name{extract_info}
\alias{extract_info}
\title{Extract Information from Line}
\usage{
extract_info(line, type)
}
\arguments{
\item{line}{A character string representing a line from a GEDCOM file.}

\item{type}{A character string representing the type of information to extract.}
}
\value{
A character string with the extracted information.
}
\description{
Extracts the relevant information from a GEDCOM line based on the specified type.
The function uses regular expressions to locate and return the desired data.
}
\keyword{internal}
