% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famSizeCal.R
\name{famSizeCal}
\alias{famSizeCal}
\title{allGens
A function to calculate the total number of individuals in a pedigree given parameters. This is a supporting function for function \code{SimPed}}
\usage{
famSizeCal(kpc, Ngen, marR)
}
\arguments{
\item{kpc}{kids per couple}

\item{Ngen}{number of generations}

\item{marR}{Mating rate.}
}
\value{
Returns a numeric value indicating the pedigree size.
}
\description{
allGens
A function to calculate the total number of individuals in a pedigree given parameters. This is a supporting function for function \code{SimPed}
}
