% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePedigree.R
\name{findBiggest}
\alias{findBiggest}
\title{This function finds the biggest families in a pedigree. It is supposed to be used internally by the \code{summarize_pedigree} function.}
\usage{
findBiggest(foo_summary_dt, nbiggest, n_foo)
}
\arguments{
\item{foo_summary_dt}{A data.table containing the summary statistics.}

\item{nbiggest}{An integer specifying the number of biggest families to find.}

\item{n_foo}{An integer specifying the number of individuals in the summary.}
}
\description{
This function finds the biggest families in a pedigree. It is supposed to be used internally by the \code{summarize_pedigree} function.
}
