% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePedigree.R
\name{findFounder}
\alias{findFounder}
\title{This function finds the originating member for each line in a pedigree. It is supposed to be used internally by the \code{summarize_pedigree} function.}
\usage{
findFounder(data, group_var, sort_var)
}
\arguments{
\item{data}{A data.table containing the pedigree data.}

\item{sort_var}{A character string specifying the column name to sort by.}
}
\value{
A data.table containing the originating member for each line.
}
\description{
This function finds the originating member for each line in a pedigree. It is supposed to be used internally by the \code{summarize_pedigree} function.
}
\keyword{internal}
