% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWikifamilytree.R
\name{findNeighbors}
\alias{findNeighbors}
\title{Build adjacency list (4-way neighbors)}
\usage{
findNeighbors(cell, active_keys)
}
\arguments{
\item{cell}{A data frame with columns Row and Column}
}
\value{
A character vector of neighboring cell keys
}
\description{
Build adjacency list (4-way neighbors)
}
\keyword{internal}
