% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePedigree.R
\name{findOldest}
\alias{findOldest}
\alias{findoldest}
\title{Function to find the oldest individuals in a pedigree
This function finds the oldest families in a pedigree. It is supposed to be used internally by the \code{summarize_pedigree} function.}
\usage{
findOldest(
  foo_summary_dt,
  byr = "byr",
  n_fooest = 5,
  n_foo_total = nrow(foo_summary_dt)
)
}
\arguments{
\item{foo_summary_dt}{A data.table containing the summary statistics.}

\item{byr}{Character. Optional column name for birth year. Used to determine the oldest lineages.}

\item{n_fooest}{An integer specifying the number of individuals in the summary.}

\item{n_foo_total}{An integer specifying the total number of individuals in the summary.}
}
\value{
a data.table containing the oldest families in the pedigree.
}
\description{
Function to find the oldest individuals in a pedigree
This function finds the oldest families in a pedigree. It is supposed to be used internally by the \code{summarize_pedigree} function.
}
