% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWikifamilytree.R
\name{getWikiTreeSummary}
\alias{getWikiTreeSummary}
\title{Extract Summary Text}
\usage{
getWikiTreeSummary(text)
}
\arguments{
\item{text}{A character string containing the text of a family tree in wiki format.}
}
\value{
A character string containing the summary text.
}
\description{
Extract Summary Text
}
\keyword{internal}
