% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BG.Misc-package.R
\name{ggbg_package}
\alias{ggbg_package}
\alias{ggbg-package}
\title{Behavior Genetic Graphing in R}
\description{
The classic behavior genetic model uses data from twins to decompose sources of phenotypic variation into the broad categories of genetic (a^2; Additive), environmental (c^2; Shared), and unique (e^2; Non-Shared) variance. The primary aim of this research is to describe the relative influence of genes and the environment on virtually every phenotype. It is used across a range of disciplines from psychology to economics.

The \code{BG.Misc} package extends the popular \code{\link[=ggplot]{ggplot2}} package. ggplot2 implements Wilkinsons
"The Grammar of Graphics", and, makes provision for a highly methodical construction process for the development
of meaningful (graphical) data representations. Of course, the above book by Wilkinson outlines the \emph{theory},
whilst Hadley Wickhams \code{\link[=ggplot]{ggplot2}} implementation is where much of the magic happens,
and, an ideal base-platform for the \code{ggbg} package.

In this document, some of the main features are highlighted, however, current examples (and corresponding outputs)
can be viewed at https://github.com/smasongarrison/BG.Misc
}
\section{\code{ggbg} Constructor}{

Plots in \code{ggbg} are instigated via the default constructor: \code{ggbg(...)},
for additional information, click \link[=ggbg]{HERE}:
}

\section{\code{ggbg} Modification to Required Aesthetics}{

}

\references{
To cite this package, please use the following:

XXXXXXXXXXXX

A bibtex entry can be obtained by executing the following command:
\code{citation('ggbg')}
}
\author{
S. Mason Garrison
}
