% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{hazard}
\alias{hazard}
\title{Simulated pedigree with two extended families and an age-related hazard}
\format{
A data frame with 43 rows and 14 variables
}
\usage{
data(hazard)
}
\description{
A dataset simulated to have an age-related hazard.
There are two extended families that are sampled from the same population.
}
\details{
The variables are as follows:

\itemize{
  \item FamID.  ID of the extended family
  \item ID.  Person identification variable
  \item sex.  Sex of the ID: 1 is female; 0 is male
  \item dadID.  ID of the father
  \item momID.  ID of the mother
  \item affected.  logical.  Whether the person is affected or not
  \item DA1.  Binary variable signifying the meaninglessness of life
  \item DA2.  Binary variable signifying the fundamental unknowability of existence
  \item birthYr.  Birth year for person
  \item onsetYr.  Year of onset for person
  \item deathYr.  Death year for person
  \item available.  logical.  Whether 
  \item Gen.  Generation of the person
  \item proband.  logical.  Whether the person is a proband or not
}
}
\keyword{datasets}
