% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{inbreeding}
\alias{inbreeding}
\title{Artificial pedigree data on seven families with inbreeding}
\format{
A data frame (and ped object) with 113 rows and 7 variables
}
\usage{
data(inbreeding)
}
\description{
A dataset created purely from imagination that includes several types of inbreeding.
Different kinds of inbreeding occur in each extended family.
}
\details{
The variables are as follows:

\itemize{
  \item ID.  Person identification variable
  \item sex.  Sex of the ID: 1 is female; 0 is male
  \item dadID.  ID of the father
  \item momID.  ID of the mother
  \item FamID.  ID of the extended family
  \item Gen.  Generation of the person
  \item proband.  Always FALSE
}
}
\keyword{datasets}
