% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLinks.R
\name{initialize_empty_df}
\alias{initialize_empty_df}
\title{initialize_empty_df}
\usage{
initialize_empty_df(relNames)
}
\arguments{
\item{relNames}{A vector of column names to be included in the data frame.}
}
\value{
An empty data frame with specified column names.
}
\description{
This function initializes an empty data frame with specified column names.
}
\keyword{internal}
