% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGedcom.R
\name{initializeRecord}
\alias{initializeRecord}
\title{Initialize an Empty Individual Record}
\usage{
initializeRecord(all_var_names)
}
\arguments{
\item{all_var_names}{A character vector of variable names.}
}
\value{
A named list representing an empty individual record.
}
\description{
Creates a named list with all GEDCOM fields set to NA.
}
