% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPedigree.R
\name{isChild}
\alias{isChild}
\title{Determine isChild Status, isChild is the 'S' matrix from RAM}
\usage{
isChild(isChild_method, ped)
}
\arguments{
\item{isChild_method}{method to determine isChild status}

\item{ped}{pedigree data frame}
}
\value{
isChild 'S' matrix
}
\description{
Determine isChild Status, isChild is the 'S' matrix from RAM
}
