% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPedigree.R
\name{makeLongTree}
\alias{makeLongTree}
\title{Make Long Tree}
\usage{
makeLongTree(tree_df, cols_to_pivot)
}
\arguments{
\item{tree_df}{A data frame containing the tree structure.}

\item{cols_to_pivot}{A character vector of column names to pivot.}
}
\value{
A long data frame containing the tree structure.
}
\description{
Make Long Tree
}
\keyword{internal}
