% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGedcom.R
\name{mapFAMC2parents}
\alias{mapFAMC2parents}
\title{Assign momID and dadID based on family mapping}
\usage{
mapFAMC2parents(df_temp, family_to_parents)
}
\arguments{
\item{df_temp}{A data frame containing individual information.}

\item{family_to_parents}{A list mapping family IDs to parent IDs.}
}
\value{
A data frame with added momID and dad_ID columns.
}
\description{
This function assigns mother and father IDs to individuals in the data frame
based on the mapping of family IDs to parent IDs.
}
\keyword{internal}
