% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGedcom.R
\name{mapFAMS2parents}
\alias{mapFAMS2parents}
\title{Create a Mapping from Family IDs to Parent IDs}
\usage{
mapFAMS2parents(df_temp)
}
\arguments{
\item{df_temp}{A data frame produced by \code{readGedcom()}.}
}
\value{
A list mapping family IDs to parent information.
}
\description{
This function scans the data frame and creates a mapping of family IDs
to the corresponding parent IDs.
}
