% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPedigree.R
\name{matchMembers}
\alias{matchMembers}
\title{Match Members}
\usage{
matchMembers(text)
}
\arguments{
\item{text}{A character string containing the text of a family tree in wiki format.}
}
\value{
A data frame containing information about the members of the family tree.
}
\description{
Match Members
}
\keyword{internal}
