% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{Null}
\alias{Null}
\title{Compute the null space of a matrix}
\usage{
Null(M)
}
\arguments{
\item{M}{a matrix of which the null space is desired}
}
\description{
Compute the null space of a matrix
}
\details{
The method uses the QR factorization to determine a basis for the null
space of a matrix.  This is sometimes also called the orthogonal
complement of a matrix.  As implemented, this function is identical
to the function of the same name in the MASS package.
}
\keyword{internal}
