% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpGeneric.R
\name{null2NA}
\alias{null2NA}
\alias{nullToNA}
\title{nullToNA}
\usage{
null2NA(x)

nullToNA(x)
}
\arguments{
\item{x}{vector of any length}
}
\value{
replaces null values in a vector to NA
}
\description{
nullToNA
}
\keyword{internal}
