% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPedigree.R
\name{parseRelationships}
\alias{parseRelationships}
\title{infer relationship from tree template}
\usage{
parseRelationships(tree_long)
}
\arguments{
\item{tree_long}{A data frame containing the tree structure in long format.}
}
\value{
A data frame containing the relationships between family members.
}
\description{
infer relationship from tree template
}
\keyword{internal}
