% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWikifamilytree.R
\name{parseTreeRelationships}
\alias{parseTreeRelationships}
\title{infer relationship from tree template}
\usage{
parseTreeRelationships(tree_long, tree_paths = NULL)
}
\arguments{
\item{tree_long}{A data frame containing the tree structure in long format.}

\item{tree_paths}{Optional. traceTreePaths output. If NULL, it will be calculated.}
}
\value{
A data frame containing the relationships between family members.
}
\description{
infer relationship from tree template
}
\keyword{internal}
