% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPedigree.R
\name{PlotPedigree}
\alias{PlotPedigree}
\title{PlotPedigree
A wrapped function to plot simulated pedigree from function \code{SimPed}. This function require the installation of package \code{kinship2}.}
\usage{
PlotPedigree(ped, cex = 0.5)
}
\arguments{
\item{ped}{The simulated pedigree data.frame from function \code{SimPed}. Or a pedigree dataframe with the same colnames as the dataframe simulated from function \code{SimPed}.}

\item{cex}{The font size of the IDs for each individual in the plot.}
}
\value{
A plot of the simulated pedigree
}
\description{
PlotPedigree
A wrapped function to plot simulated pedigree from function \code{SimPed}. This function require the installation of package \code{kinship2}.
}
