% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWikifamilytree.R
\name{populateParents}
\alias{populateParents}
\title{Assign Parent}
\usage{
populateParents(df, child, parent)
}
\arguments{
\item{df}{A data frame containing the relationships.}

\item{child}{The ID of the child.}

\item{parent}{The ID of the parent.}
}
\value{
A data frame with updated parent information.
}
\description{
Assign Parent
}
\keyword{internal}
