% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGedcom.R
\name{processParents}
\alias{processParents}
\title{Process Parents Information from GEDCOM Data}
\usage{
processParents(df_temp, datasource)
}
\arguments{
\item{df_temp}{A data frame produced by \code{readGedcom()}.}

\item{datasource}{Character string indicating the data source ("gedcom" or "wiki").}
}
\value{
The updated data frame with parent IDs added.
}
\description{
Adds parent IDs to the individuals based on family relationship data.
}
