% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDelimitedData.R
\name{readDelimitedData}
\alias{readDelimitedData}
\title{Load Delimited Data File}
\usage{
readDelimitedData(
  path = "Z:/Data/2024/Smith_Burt_143000_Demographic_Export.txt",
  var_nrows,
  delim = "|",
  as.is = TRUE
)
}
\arguments{
\item{path}{A character string specifying the file path to the data file. Default is 'Z:/Data/2024/Smith_Burt_143000_Demographic_Export.txt'.}

\item{var_nrows}{An integer specifying the number of rows to read from the file.}

\item{delim}{A character string specifying the delimiter used in the data file. Default is '|'. Accepts '|' or ','.}

\item{as.is}{Logical. If TRUE, keeps character columns as strings without converting to factors. Default is TRUE.}
}
\value{
A data frame containing the parsed data.
}
\description{
This function reads a text file containing delimited data and converts it into a data frame.
It supports multiple delimiters, including pipe ('|') and comma (',').
}
\examples{
# Example usage
df <- loadDelimitedData(path = 'data.txt', var_nrows = 100, delim = '|', as.is = TRUE)

}
