% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWikifamilytree.R
\name{readWikifamilytree}
\alias{readWikifamilytree}
\title{Read Wiki Family Tree}
\usage{
readWikifamilytree(text = NULL, verbose = FALSE, file_path = NULL, ...)
}
\arguments{
\item{text}{A character string containing the text of a family tree in wiki format.}

\item{verbose}{A logical value indicating whether to print messages.}

\item{file_path}{The path to the file containing the family tree.}

\item{...}{Additional arguments (not used).}
}
\value{
A list containing the summary, members, structure, and relationships of the family tree.
}
\description{
Read Wiki Family Tree
}
