% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatePedigree.R
\name{recodeSex}
\alias{recodeSex}
\title{recodeSex
An internal utility function for \code{plotPedigree} to handle the recoding of the 'sex' variable in a pedigree dataframe.}
\usage{
recodeSex(ped, code_male)
}
\arguments{
\item{ped}{A pedigree data.frame containing the relevant columns, including 'sex'.}

\item{code_male}{This optional input allows you to indicate what value in the sex variable codes for male. Will be recoded as "M" (Male). If \code{NULL}, no recoding is performed.}
}
\value{
A modified version of the input data.frame \code{ped}, containing an additional or modified 'sex_recode' column where the 'sex' values are recoded according to \code{code_male}. NA values in the 'sex' column are preserved.
}
\description{
recodeSex
An internal utility function for \code{plotPedigree} to handle the recoding of the 'sex' variable in a pedigree dataframe.
}
\seealso{
\code{\link{plotPedigree}}
}
\keyword{internal}
