% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeIDs.R
\name{relabelMatrix}
\alias{relabelMatrix}
\title{Relabel Matrix Dimnames Using an ID Map}
\usage{
relabelMatrix(mat, id_map)
}
\arguments{
\item{mat}{A matrix or Matrix object whose dimnames are (character
representations of) numeric surrogate IDs.}

\item{id_map}{A mapping data.frame from \code{\link{idMap}}, with columns
\code{original_id} and \code{numeric_id}.}
}
\value{
The same matrix with dimnames replaced by original IDs.
}
\description{
Translates the row and column names of a matrix (or Matrix) from numeric
surrogate IDs back to original IDs using a mapping table. This is used
internally to label relatedness and adjacency matrices with the user's
original IDs after computation is done with integer surrogates.
}
\examples{
m <- matrix(1:9, 3, 3, dimnames = list(1:3, 1:3))
map <- data.frame(original_id = c("A", "B", "C"), numeric_id = 1:3)
relabelMatrix(m, map)

}
\keyword{internal}
