% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{related_coef}
\alias{related_coef}
\title{related_coef (Deprecated)}
\usage{
related_coef(...)
}
\arguments{
\item{...}{Arguments to be passed to `calculateRelatedness`.}
}
\value{
The same result as calling `calculateRelatedness`.
}
\description{
When calling this function, a warning will be issued about its deprecation.
}
\details{
This function is a wrapper around the new `calculateRelatedness` function.
`related_coef` has been deprecated, and it's advised to use `calculateRelatedness` directly.
}
\examples{
\dontrun{
# This is an example of the deprecated function:
related_coef(...)
# It is recommended to use:
calculateRelatedness(...)
}
}
\seealso{
\code{\link{calculateRelatedness}} for the updated function.
}
\keyword{deprecated}
