% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{relatedness}
\alias{relatedness}
\title{Calculate Relatedness Coefficient}
\usage{
relatedness(cor_obs, ace_A = 0.9, ace_C = 0, shared_C = 0)
}
\arguments{
\item{cor_obs}{Numeric. Observed correlation between the two groups Must be a value between -1 and 1.}

\item{ace_A}{Numeric. Proportion of variance attributable to additive genetic variance. Must be a value between 0 and 1. Default is 0.9.}

\item{ace_C}{Numeric. Proportion of variance attributable to shared environmental variance. Must be a value between 0 and 1. Default is 0.}

\item{shared_C}{Numeric. Proportion of shared environment shared between the two individuals. Must be a value between 0 and 1. Default is 0.}
}
\value{
Numeric. Calculated relatedness coefficient `est_r`.
}
\description{
This function calculates the relatedness coefficient between two groups based on the observed correlation between their additive genetic variance and shared environmental variance.
}
\examples{
# Using the ACE framework, we can calculate the relatedness between two
# individuals based on the observed correlation between their additive genetic
# variance, shared environmental variance, and proportion of shared environment.

relatedness(cor_obs = 0.5, ace_A = 0.9, ace_C = 0, shared_C = 0)

}
