% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIDs.R
\name{repairIDs}
\alias{repairIDs}
\title{Repair Missing IDs}
\usage{
repairIDs(ped, verbose = FALSE)
}
\arguments{
\item{ped}{A dataframe representing the pedigree data with columns `ID`, `dadID`, and `momID`.}

\item{verbose}{A logical flag indicating whether to print progress and validation messages to the console.}
}
\value{
A corrected pedigree
}
\description{
This function repairs missing IDs in a pedigree.
}
