% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpNames.R
\name{restorePedColnames}
\alias{restorePedColnames}
\title{Restore Original Column Names in a Pedigree Dataframe}
\usage{
restorePedColnames(
  ped,
  famID = "famID",
  personID = "ID",
  momID = "momID",
  dadID = "dadID",
  gen = "gen",
  patID = "patID",
  matID = "matID",
  spID = "spID",
  twinID = "twinID",
  zygosity = "zygosity",
  sex = "sex"
)
}
\arguments{
\item{ped}{A pedigree dataframe with standardized column names.}

\item{famID}{The original name for the family ID column. Default is "fam
ID".}

\item{personID}{The original name for the person ID column. Default is "ID".}

\item{momID}{The original name for the mother ID column. Default is "momID".}

\item{dadID}{The original name for the father ID column. Default is "dadID".}

\item{gen}{The original name for the generation column. Default is "gen".}

\item{patID}{The original name for the paternal ID column. Default is "patID".}

\item{matID}{The original name for the maternal ID column. Default is "matID".}

\item{spID}{The original name for the spouse ID column. Default is "spID".}

\item{twinID}{The original name for the twin ID column. Default is "twinID".}

\item{zygosity}{The original name for the zygosity column. Default is "zygosity".}

\item{sex}{The original name for the sex column. Default is "sex".}
}
\value{
A pedigree dataframe with restored original column names.
}
\description{
This function restores the original column names of a pedigree dataframe
based on user-specified names. It is useful for reverting standardized column
names back to their original names after processing.
}
