% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpGeneric.R
\name{SimPed}
\alias{SimPed}
\title{SimPed (Deprecated)}
\usage{
SimPed(...)
}
\arguments{
\item{...}{Arguments to be passed to `simulatePedigree`.}
}
\value{
The same result as calling `simulatePedigree`.
}
\description{
When calling this function, a warning will be issued about its deprecation.
}
\details{
This function is a wrapper around the new `simulatePedigree` function.
`SimPed` has been deprecated, and it's advised to use `simulatePedigree` directly.
}
\examples{
\dontrun{
# This is an example of the deprecated function:
SimPed(...)
# It is recommended to use:
simulatePedigree(...)
}
}
\seealso{
\code{\link{simulatePedigree}} for the updated function.
}
\keyword{deprecated}
