% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFamilySize.R
\name{sizeAllGens}
\alias{sizeAllGens}
\title{sizeAllGens
An internal supporting function for \code{simulatePedigree}.}
\usage{
sizeAllGens(kpc, Ngen, marR)
}
\arguments{
\item{kpc}{Number of kids per couple (integer >= 2).}

\item{Ngen}{Number of generations (integer >= 1).}

\item{marR}{Mating rate (numeric value ranging from 0 to 1).}
}
\value{
Returns a vector including the number of individuals in every generation.
}
\description{
sizeAllGens
An internal supporting function for \code{simulatePedigree}.
}
