% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{standardize_colnames}
\alias{standardize_colnames}
\title{Standardize Column Names in a Dataframe (Internal)}
\usage{
standardize_colnames(df)
}
\arguments{
\item{df}{A dataframe whose column names need to be standardized.}
}
\value{
A dataframe with standardized column names.
}
\description{
This internal function standardizes the column names of a given dataframe.
It is useful for ensuring that dataframes with varying column names can be processed by functions
expecting specific column names.
}
\keyword{internal}
