% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPedigree.R
\name{standardizeColnames}
\alias{standardizeColnames}
\title{Standardize Column Names in a Dataframe (Internal)}
\usage{
standardizeColnames(df, verbose = FALSE, mapping = list())
}
\arguments{
\item{df}{A dataframe whose column names need to be standardized.}

\item{verbose}{A logical indicating whether to print progress messages.}

\item{mapping}{A list of mapping options for customizing the renaming process.}
}
\value{
A dataframe with standardized column names.
}
\description{
This internal function standardizes the column names of a given dataframe.
It utilizes regular expressions and the `tolower()` function to match column names
against a list of predefined standard names. The approach is case-insensitive and
allows for flexible matching of column names.
}
