% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePedigree.R
\name{summarizeFamilies}
\alias{summarizeFamilies}
\title{Summarize the families in a pedigree}
\usage{
summarizeFamilies(
  ped,
  famID = "famID",
  personID = "ID",
  momID = "momID",
  dadID = "dadID",
  matID = "matID",
  patID = "patID",
  byr = NULL,
  nbiggest = 5,
  noldest = 5,
  skipvar = NULL
)
}
\arguments{
\item{ped}{a pedigree dataset.  Needs ID, momID, and dadID columns}

\item{famID}{character.  Name of the column to be created in ped for the family ID variable}

\item{personID}{character.  Name of the column in ped for the person ID variable}

\item{momID}{character.  Name of the column in ped for the mother ID variable}

\item{dadID}{character.  Name of the column in ped for the father ID variable}

\item{matID}{Character.  Maternal line ID variable to be created and added to the pedigree}

\item{patID}{Character.  Paternal line ID variable to be created and added to the pedigree}

\item{byr}{The column name for birth year.}

\item{nbiggest}{The number of biggest lines to return.}

\item{noldest}{The number of oldest lines to return.}

\item{skipvar}{A character vector of variables to skip when calculating summary statistics.}
}
\description{
Summarize the families in a pedigree
}
\seealso{
[summarizePedigrees ()]
}
