% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePedigree.R
\name{summarizeFounder}
\alias{summarizeFounder}
\title{Function to summarize the originating members for each line}
\usage{
summarizeFounder(ped_dt, group_var, sort_var, foo_summary_dt, verbose)
}
\arguments{
\item{sort_var}{A character string specifying the column name to sort by.}

\item{verbose}{Logical, if TRUE, print progress messages.}
}
\description{
This function summarizes the originating members for each line in a pedigree.
It is supposed to be used internally by the \code{summarize_pedigree} function.
}
\keyword{internal}
