% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePedigree.R
\name{summarizeOldest}
\alias{summarizeOldest}
\title{Function to summarize the oldest individuals in a pedigree}
\usage{
summarizeOldest(
  byr = NULL,
  n_oldest = 5,
  n_families = NULL,
  type = NULL,
  verbose = FALSE,
  output,
  family_summary_dt = NULL,
  n_mothers = NULL,
  maternal_summary_dt = NULL,
  n_fathers = NULL,
  paternal_summary_dt = NULL
)
}
\arguments{
\item{byr}{Character. Optional column name for birth year. Used to determine the oldest lineages.}

\item{n_oldest}{Integer. Number of oldest lineages to return (sorted by birth year).}

\item{type}{Character vector. Specifies which summaries to compute.
Options: `"fathers"`, `"mothers"`, `"families"`. Default includes all three.}

\item{verbose}{Logical, if TRUE, print progress messages.}
}
\value{
A data.table containing the summary statistics for all numeric variables.
}
\description{
Function to summarize the oldest individuals in a pedigree
}
\keyword{internal}
