% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePedigree.R
\name{summarizePedigree}
\alias{summarizePedigree}
\title{Summarize Pedigree Data}
\usage{
summarizePedigree(
  ped,
  famID = "famID",
  personID = "ID",
  momID = "momID",
  dadID = "dadID",
  matID = "matID",
  patID = "patID",
  yrb = NULL,
  nbiggest = 5,
  noldest = 5
)
}
\arguments{
\item{personID}{A character string indicating the column name for the person ID variable.}

\item{momID}{A character string indicating the column name for the mother ID variable.}

\item{dadID}{A character string indicating the column name for the father ID variable.}

\item{pedigree_data}{A data frame containing the pedigree data.}
}
\value{
A list containing summary statistics for family, maternal, and paternal lines, as well as the 5 oldest and biggest lines.
}
\description{
This function summarizes pedigree data, including calculating summary statistics for all numeric variables,
and finding the originating member for each family, maternal, and paternal line.
}
