% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePedigree.R
\name{summarizePedigrees}
\alias{summarizePedigrees}
\title{Summarize Pedigree Data}
\usage{
summarizePedigrees(
  ped,
  famID = "famID",
  personID = "ID",
  momID = "momID",
  dadID = "dadID",
  matID = "matID",
  patID = "patID",
  byr = NULL,
  type = c("fathers", "mothers", "families"),
  nbiggest = 5,
  noldest = 5
)
}
\arguments{
\item{ped}{a pedigree dataset.  Needs ID, momID, and dadID columns}

\item{famID}{character.  Name of the column to be created in ped for the family ID variable}

\item{personID}{character.  Name of the column in ped for the person ID variable}

\item{momID}{character.  Name of the column in ped for the mother ID variable}

\item{dadID}{character.  Name of the column in ped for the father ID variable}

\item{matID}{Character.  Maternal line ID variable to be created and added to the pedigree}

\item{patID}{Character.  Paternal line ID variable to be created and added to the pedigree}

\item{byr}{The column name for birth year.}

\item{type}{The type of summary statistics to calculate. Options are "fathers", "mothers", and "families".}

\item{nbiggest}{The number of biggest lines to return.}

\item{noldest}{The number of oldest lines to return.}
}
\value{
A list containing summary statistics for family, maternal, and paternal lines, as well as the 5 oldest and biggest lines.
}
\description{
This function summarizes pedigree data, including calculating summary statistics for all numeric variables,
and finding the originating member for each family, maternal, and paternal line.
}
