% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWikifamilytree.R
\name{traceTreePaths}
\alias{traceTreePaths}
\title{Trace paths between individuals in a family tree grid}
\usage{
traceTreePaths(tree_long, deduplicate = TRUE)
}
\arguments{
\item{tree_long}{A data.frame with columns: Row, Column, Value, id}

\item{deduplicate}{Logical, if TRUE, will remove duplicate paths}
}
\value{
A data.frame with columns: from_id, to_id, direction, path_length, intermediates
}
\description{
Trace paths between individuals in a family tree grid
}
