% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpGeneric.R
\name{tryNA}
\alias{tryNA}
\alias{try_na}
\title{modified tryCatch function}
\usage{
tryNA(x)

try_na(x)
}
\arguments{
\item{x}{vector of any length}
}
\value{
Fuses the nullToNA function with efunc
}
\description{
modified tryCatch function
}
\keyword{internal}
