% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famSizeCal.R
\name{twinImpute}
\alias{twinImpute}
\title{twinImpute
A function to impute twins in the simulated pedigree \code{data.frame}.
Twins can be imputed by specifying their IDs or by specifying the generation the twin should be imputed.
This is a supplementary function for \code{simulatePedigree}.}
\usage{
twinImpute(ped, ID_twin1 = NA_integer_, ID_twin2 = NA_integer_, gen_twin = 2)
}
\arguments{
\item{ped}{A \code{data.frame} in the same format as the output of \code{simulatePedigree}.}

\item{ID_twin1}{A vector of \code{ID} of the first twin.}

\item{ID_twin2}{A vector of \code{ID} of the second twin.}

\item{gen_twin}{A vector of \code{generation} of the twin to be imputed.}
}
\value{
Returns a \code{data.frame} with MZ twins infomration added as a new column.
}
\description{
twinImpute
A function to impute twins in the simulated pedigree \code{data.frame}.
Twins can be imputed by specifying their IDs or by specifying the generation the twin should be imputed.
This is a supplementary function for \code{simulatePedigree}.
}
