% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLinks.R
\name{validate_and_convert_matrix}
\alias{validate_and_convert_matrix}
\title{validate_and_convert_matrix}
\usage{
validate_and_convert_matrix(
  mat,
  name,
  ensure_symmetric = FALSE,
  force_binary = FALSE
)
}
\arguments{
\item{mat}{The matrix to be validated and converted.}

\item{name}{The name of the matrix for error messages.}

\item{ensure_symmetric}{Logical indicating whether to ensure the matrix is symmetric.}

\item{force_binary}{Logical indicating whether to force the matrix to be binary.}
}
\value{
The validated and converted matrix.
}
\description{
This function validates and converts a matrix to a specific format.
}
