% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatePedigree.R
\name{validatePedigree}
\alias{validatePedigree}
\title{Validate Pedigrees}
\usage{
validatePedigree(
  ped,
  verbose = FALSE,
  check_sex = TRUE,
  check_parents = TRUE,
  check_id = TRUE
)
}
\arguments{
\item{ped}{A pedigree object}

\item{verbose}{Logical. Indicates whether to print progress messages}

\item{check_sex}{Logical. Indicates that sex should be validated}

\item{check_parents}{Logical. Indicates that parents should be validated}

\item{check_id}{Logical. Indicates that IDs should be validated}

\item{...}{Additional arguments to be passed to \code{\link{validatePedigree}}}
}
\value{
A logical indicating whether the pedigree is valid and a list of warnings and ids of potentially invalid relationships
}
\description{
This function validates pedigrees based on several criteria.
}
