% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{vech}
\alias{vech}
\title{Create the half-vectorization of a matrix}
\usage{
vech(x)
}
\arguments{
\item{x}{a matrix, the half-vectorization of which is desired}
}
\description{
Create the half-vectorization of a matrix
}
\details{
Returns the vector of the lower triangle of a matrix, including the diagonal.
The upper triangle is ignored with no checking that the provided matrix
is symmetric.
}
