% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addParentalChain.R
\name{addPaternalChain}
\alias{addPaternalChain}
\title{This is a convenience wrapper around [addParentalChain()] with
`component = "dadID"`.}
\usage{
addPaternalChain(
  ped,
  personID = "personID",
  dadID = "dadID",
  momID = "momID",
  chain_col = "dadID_chain",
  chain_string_col = "dadID_chain_string",
  collapse = "|",
  traversal_direction = "in"
)
}
\arguments{
\item{ped}{A pedigree data frame.}

\item{personID}{Character string giving the name of the column containing
individual IDs.}

\item{dadID}{Character string giving the name of the column containing
paternal IDs.}

\item{momID}{Character string giving the name of the column containing
maternal IDs.}

\item{chain_col}{Character string giving the name of the output list-column
that will contain the ordered parental ancestor chain for each individual.}

\item{chain_string_col}{Character string giving the name of the output
character column that will contain the collapsed parental ancestor chain.}

\item{collapse}{Character string used to collapse ancestor IDs into
`chain_string_col`.}

\item{traversal_direction}{Character giving the mode of transversion: defaults to "in". If set to out, will procedure a list of descendants.}
}
\value{
A data frame with two added columns:
  \describe{
    \item{`chain_col`}{A list-column containing the ordered paternal ancestor
    chain for each individual.}
    \item{`chain_string_col`}{A character column containing the collapsed
    paternal ancestor chain, or `NA_character_` when no paternal ancestors are
    found.}
  }
}
\description{
This is a convenience wrapper around [addParentalChain()] with
`component = "dadID"`.
}
