% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkParents.R
\name{addRowlessParents}
\alias{addRowlessParents}
\title{Add addRowlessParents}
\usage{
addRowlessParents(ped, verbose, validation_results)
}
\arguments{
\item{ped}{A dataframe representing the pedigree data with columns 'ID', 'dadID', and 'momID'.}

\item{verbose}{A logical flag indicating whether to print progress and validation messages to the console.}

\item{validation_results}{validation results}
}
\description{
This function adds parents who appear in momID or dadID but are missing from ID
}
