% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmxPedigrees.R
\name{alignPhenToMatrix}
\alias{alignPhenToMatrix}
\title{Align Phenotype Vector to Matrix Format for OpenMx}
\usage{
alignPhenToMatrix(ped, phenotype, keep_ids, id_col = "ID")
}
\arguments{
\item{ped}{A data frame representing the pedigree, containing at least the columns specified by \code{phenotype} and \code{id_col}.}

\item{phenotype}{A character string specifying the column name in \code{ped} that
contains the phenotype values to be aligned.
@param keep_ids A vector of IDs for which the phenotype values should be extracted and aligned. These IDs should correspond to the values in the \code{id_col} of \code{ped}.
@param id_col A character string specifying the column name in \code{ped} that contains the individual IDs. Default is "ID".}
}
\description{
This function takes a pedigree data frame, a specified phenotype column, and a vector of IDs to keep, and returns a matrix formatted for use in OpenMx models. The resulting matrix has one row and columns corresponding to the specified IDs, with values taken from the phenotype column of the pedigree.
}
