% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{ASOIAF}
\alias{ASOIAF}
\title{A song of ice and fire pedigree data}
\format{
A data frame with 501 observations
}
\usage{
data(ASOIAF)
}
\description{
A dataset created from the Song of Ice and Fire series by George R. R. Martin. Core data is from the [Westeros.org forum](https://asoiaf.westeros.org/index.php?/topic/88863-all-the-family-trees/).
}
\details{
The variables are as follows:
\itemize{
  \item \code{id}:  Person identification variable
  \item \code{momID}:  ID of the mother
  \item \code{dadID}:  ID of the father
  \item \code{name}:  Name of the person
  \item \code{sex}: Biological sex
  }
}
\keyword{datasets}
