% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGmisc-package.R
\docType{package}
\name{BGmisc-package}
\alias{BGmisc}
\alias{BGmisc-package}
\title{BGmisc: An R Package for Extended Behavior Genetics Analysis}
\description{
Provides functions for behavior genetics analysis, including variance component model identification [Hunter et al. (2021) \doi{10.1007/s10519-021-10055-x}], calculation of relatedness coefficients using path-tracing methods [Wright (1922) \doi{10.1086/279872}; McArdle & McDonald (1984) \doi{10.1111/j.2044-8317.1984.tb00802.x}], inference of relatedness, pedigree conversion, and simulation of multi-generational family data [Lyu et al. (2025) \doi{10.1007/s10519-025-10225-1}]. For a full overview, see [Garrison et al. (2024) \doi{10.21105/joss.06203}]. For a big data application see [Burt et al. (2025) \doi{ 10.1016/j.ebiom.2025.105911}].
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/R-Computing-Lab/BGmisc/}
  \item \url{https://r-computing-lab.github.io/BGmisc/}
  \item Report bugs at \url{https://github.com/R-Computing-Lab/BGmisc/issues}
}

}
\author{
\strong{Maintainer}: S. Mason Garrison \email{garrissm@wfu.edu} (\href{https://orcid.org/0000-0002-4804-6003}{ORCID})

Authors:
\itemize{
  \item Michael D. Hunter (\href{https://orcid.org/0000-0002-3651-6709}{ORCID})
  \item Xuanyu Lyu (\href{https://orcid.org/0000-0002-2841-5529}{ORCID})
  \item Jonathan D. Trattner \email{code@jdtrat.com} (\href{https://orcid.org/0000-0002-1097-7603}{ORCID})
  \item S. Alexandra Burt (\href{https://orcid.org/0000-0001-5538-7431}{ORCID})
}

Other contributors:
\itemize{
  \item Rachel N. Good [contributor]
}

}
\keyword{internal}
