% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGmisc-package.R
\docType{package}
\name{BGmisc-package}
\alias{BGmisc}
\alias{BGmisc-package}
\title{BGmisc: An R Package for Extended Behavior Genetics Analysis}
\description{
The BGmisc R package offers a comprehensive suite of functions tailored for extended behavior genetics analysis, including model identification, calculating relatedness, pedigree conversion, pedigree simulation, and more.
}
\author{
\strong{Maintainer}: S. Mason Garrison \email{garrissm@wfu.edu} (\href{https://orcid.org/0000-0002-4804-6003}{ORCID})

Authors:
\itemize{
  \item Michael D. Hunter (\href{https://orcid.org/0000-0002-3651-6709}{ORCID})
  \item Xuanyu Lyu (\href{https://orcid.org/0000-0002-2841-5529}{ORCID})
  \item Rachel N. Good (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})
  \item Jonathan D. Trattner (\href{https://orcid.org/0000-0002-1097-7603}{ORCID})
  \item S. Alexandra Burt (\href{https://orcid.org/0000-0001-5538-7431}{ORCID})
}

}
\keyword{internal}
