% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGmisc-package.R
\docType{package}
\name{BGmisc_package}
\alias{BGmisc_package}
\alias{BGmisc-package}
\title{Behavior Genetic Miscellaneous functions in R}
\description{
This collection contains functions for behavior genetic
  modeling. These functions include model identification, calculating
  relatedness, and various others (e.g.  Hunter, Garrison, et al, 2021
  <doi:10.1007/s10519-019-09973-8>). See the package vignette for details: \cr\cr
}
\author{
S. Mason Garrison, Michael D. Hunter, Xuanyu Lyu, Jonathan D. Trattner, and
S. Alexandra Burt
}
