% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmxPedigrees.R
\name{buildFamilyGroups}
\alias{buildFamilyGroups}
\title{Build family group models}
\usage{
buildFamilyGroups(
  dat,
  obs_ids,
  Addmat = NULL,
  Nucmat = NULL,
  Extmat = NULL,
  Mtdmat = NULL,
  Amimat = NULL,
  Dmgmat = NULL,
  prefix = "fam"
)
}
\arguments{
\item{dat}{A data frame where each row represents a family group and columns correspond to observed variables.}

\item{obs_ids}{A character vector of individual IDs corresponding to the columns of \code{dat}
and the rows/columns of the relatedness matrices.}

\item{Addmat}{Additive genetic relatedness matrix.}

\item{Nucmat}{Nuclear family shared environment relatedness matrix.}

\item{Extmat}{Extended family shared environment relatedness matrix.}

\item{Mtdmat}{Mitochondrial genetic relatedness matrix.}

\item{Amimat}{Additive by mitochondrial interaction relatedness matrix.}

\item{Dmgmat}{Dominance genetic relatedness matrix.}

\item{prefix}{A prefix for naming the family groups. Default is "fam".}
}
\value{
A list of OpenMx models for each family group.
}
\description{
This function constructs OpenMx models for multiple family groups based on
provided relatedness matrices and observed data.
}
