% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmxPedigrees.R
\name{buildOneFamilyGroup}
\alias{buildOneFamilyGroup}
\title{Build one family group model
This function constructs an OpenMx model for a single family group based on provided relatedness matrices and observed data.}
\usage{
buildOneFamilyGroup(
  group_name,
  Addmat = NULL,
  Nucmat = NULL,
  Extmat = NULL,
  Mtdmat = NULL,
  Amimat = NULL,
  Dmgmat = NULL,
  full_df_row,
  ytemp
)
}
\arguments{
\item{group_name}{Name of the family group.}

\item{Addmat}{Additive genetic relatedness matrix.}

\item{Nucmat}{Nuclear family shared environment relatedness matrix.}

\item{Extmat}{Extended family shared environment relatedness matrix.}

\item{Mtdmat}{Mitochondrial genetic relatedness matrix.}

\item{Amimat}{Additive by mitochondrial interaction relatedness matrix.}

\item{Dmgmat}{Dominance genetic relatedness matrix.}

\item{full_df_row}{A matrix representing the observed data for the family group.}

\item{ytemp}{A vector of variable names corresponding to the observed data.}
}
\value{
An OpenMx model for the specified family group.
}
\description{
Build one family group model
This function constructs an OpenMx model for a single family group based on provided relatedness matrices and observed data.
}
