% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmxPedigrees.R
\name{buildPedigreeMx}
\alias{buildPedigreeMx}
\title{Build Pedigree mxModel
This function constructs an OpenMx pedigree model by combining variance component models and family group models.}
\usage{
buildPedigreeMx(model_name, vars, group_models)
}
\arguments{
\item{model_name}{Name of the overall pedigree model.}

\item{vars}{A named vector of initial variance component values.}

\item{group_models}{A list of OpenMx models for each family group.}
}
\value{
An OpenMx pedigree model combining variance components and family groups.
}
\description{
Build Pedigree mxModel
This function constructs an OpenMx pedigree model by combining variance component models and family group models.
}
