% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCIs.R
\name{calculateCIs}
\alias{calculateCIs}
\title{Calculate Confidence Intervals for Correlation Coefficients}
\usage{
calculateCIs(tbl, rho_var, se_var, method = "raykov", conf_level = 0.95)
}
\arguments{
\item{tbl}{A data frame or tibble containing the correlation coefficient and standard error variables.}

\item{rho_var}{The name of the column in \code{tbl} that contains the correlation coefficients.}

\item{se_var}{The name of the column in \code{tbl} that contains the standard errors.}

\item{method}{The method to use for calculating the confidence intervals. Options are "raykov", "fisherz", "doubleenteredconserv", or "doubleentered".}

\item{conf_level}{The confidence level for the intervals. Default is 0.95.}
}
\value{
A modified version of \code{tbl} with additional columns for the confidence intervals and related statistics.
}
\description{
This function calculates confidence intervals for correlation coefficients using different methods.
}
\examples{
tbl <- data.frame(rho = c(0.5, 0.7, 0.3), se = c(0.1, 0.2, 0.05))
calculateCIs(tbl, rho_var = "rho", se_var = "se", method = "raykov")

}
