% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePedigree.R
\name{calculateSummaryDT}
\alias{calculateSummaryDT}
\title{Function to calculate summary statistics for all numeric variables
This function calculates summary statistics for all numeric variables in a data.table. It is supposed to be used internally by the \code{summarize_pedigree} function.}
\usage{
calculateSummaryDT(
  data,
  group_var,
  skip_var,
  five_num_summary = FALSE,
  na_rm = TRUE
)
}
\arguments{
\item{data}{A data.table containing the pedigree data.}

\item{group_var}{A character string specifying the column name of the grouping variable.}

\item{skip_var}{Character vector. Variables to exclude from summary calculations.}

\item{five_num_summary}{Logical. If `TRUE`, includes the first quartile (Q1) and third quartile (Q3) in addition to
the minimum, median, and maximum values.}

\item{na_rm}{Logical. If `TRUE`, removes `NA` values when calculating statistics.}
}
\value{
A data.table containing the summary statistics for all numeric variables.
}
\description{
Function to calculate summary statistics for all numeric variables
This function calculates summary statistics for all numeric variables in a data.table. It is supposed to be used internally by the \code{summarize_pedigree} function.
}
\keyword{internal}
