% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGedcom.R
\name{collapseNames}
\alias{collapseNames}
\title{collapse Names}
\usage{
collapseNames(verbose, df_temp)
}
\arguments{
\item{verbose}{A logical value indicating whether to print messages.}

\item{df_temp}{A data frame containing the columns to be combined.}
}
\value{
A data frame with the combined columns.
}
\description{
This function combines the `name_given` and `name_given_pieces` columns in a data frame.
}
