% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLinks.R
\name{com2links}
\alias{com2links}
\title{Take a component and turn it into kinship links}
\usage{
com2links(
  rel_pairs_file = "dataRelatedPairs.csv",
  ad_ped_matrix = NULL,
  mit_ped_matrix = mt_ped_matrix,
  mt_ped_matrix = NULL,
  cn_ped_matrix = NULL,
  pat_ped_matrix = NULL,
  mat_ped_matrix = NULL,
  mapa_id_file = "data_mapaID.csv",
  gc = TRUE,
  writetodisk = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{rel_pairs_file}{File to write related pairs to}

\item{ad_ped_matrix}{Matrix of additive genetic relatedness coefficients}

\item{mit_ped_matrix}{Matrix of mitochondrial relatedness coefficients}

\item{cn_ped_matrix}{Matrix of common nuclear relatedness coefficients}

\item{pat_ped_matrix}{Matrix of paternal relatedness coefficients}

\item{mat_ped_matrix}{Matrix of maternal relatedness coefficients}

\item{mapa_id_file}{File to write the map of parental IDs to individual IDs}

\item{gc}{logical. If TRUE, do frequent garbage collection via \code{\link{gc}} to save memory}

\item{writetodisk}{logical. If TRUE, write the related pairs to disk}

\item{verbose}{logical. If TRUE, print progress messages}

\item{...}{Additional arguments to be passed to \code{\link{com2links}}}
}
\value{
A data frame of related pairs
}
\description{
Take a component and turn it into kinship links
}
