% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPedigree.R
\name{compute_parent_adjacency}
\alias{compute_parent_adjacency}
\title{Compute Parent Adjacency Matrix with Multiple Approaches}
\usage{
compute_parent_adjacency(
  ped,
  component,
  adjacency_method = "loop",
  saveable,
  resume,
  save_path,
  verbose,
  lastComputed,
  nr,
  parList,
  lens
)
}
\arguments{
\item{ped}{Pedigree dataset with ID, momID, and dadID columns}

\item{component}{Which relatedness component to compute}

\item{saveable}{If TRUE, saves intermediate results}

\item{resume}{If TRUE, resumes from a checkpoint}

\item{save_path}{Path to save checkpoint files}

\item{verbose}{If TRUE, prints progress updates}

\item{method}{Character, method to use ("loop", "vectorized")}
}
\value{
Sparse parent adjacency matrix
}
\description{
Compute Parent Adjacency Matrix with Multiple Approaches
}
