% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmxPedigrees.R
\name{condenseMatrixSlots}
\alias{condenseMatrixSlots}
\title{Condense Matrix Slots in an OpenMx Model}
\usage{
condenseMatrixSlots(model)
}
\arguments{
\item{model}{An OpenMx model object for which to condense matrix slots. If
NULL, the function returns NULL.}
}
\value{
An OpenMx model with condensed matrix slots, or NULL if the input model
is NULL.
}
\description{
This function takes an OpenMx model and applies the \code{mxCondenseMatrixSlots} wrapper to optimize memory usage for large matrices. This can be particularly beneficial when working with large pedigree models that include multiple relatedness matrices.
}
