% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpSimulatePedigree.R,
%   R/helpSimulatePedigree_beta.R
\name{determineSex}
\alias{determineSex}
\alias{determineSex_beta}
\title{Determine Sex of Offspring}
\usage{
determineSex(idGen, sexR, code_male = "M", code_female = "F", beta = FALSE)

determineSex_beta(idGen, sexR, code_male = "M", code_female = "F")
}
\arguments{
\item{idGen}{Vector of IDs for the generation.}

\item{sexR}{Numeric value indicating the sex ratio (proportion of males).}

\item{code_male}{The value to use for males. Default is "M"}

\item{code_female}{The value to use for females. Default is "F"}

\item{beta}{logical. If TRUE, use the optimized version of the algorithm.}
}
\value{
Vector of sexes ("M" for male, "F" for female) for the offspring.
}
\description{
This internal function assigns sexes to the offspring in a generation based on the specified sex ratio.
}
