% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpPedigree.R
\name{determineSex}
\alias{determineSex}
\title{Determine Sex of Offspring}
\usage{
determineSex(idGen, sexR, code_male = "M", code_female = "F")
}
\arguments{
\item{idGen}{Vector of IDs for the generation.}

\item{sexR}{Numeric value indicating the sex ratio (proportion of males).}

\item{code_male}{The value to use for males. Default is "M"}

\item{code_female}{The value to use for females. Default is "F"}
}
\value{
Vector of sexes ("M" for male, "F" for female) for the offspring.
}
\description{
This internal function assigns sexes to the offspring in a generation based on the specified sex ratio.
}
