% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildComponent.R
\name{.computeTranspose}
\alias{.computeTranspose}
\title{Compute the transpose multiplication for the relatedness matrix}
\usage{
.computeTranspose(
  r2,
  transpose_method = "tcrossprod",
  chunk_size = 1000L,
  verbose = FALSE,
  force_symmetric = FALSE,
  config = NULL,
  checkpoint_files = NULL
)
}
\arguments{
\item{r2}{a relatedness matrix}

\item{transpose_method}{character. The method to use for computing the transpose.  Options are "tcrossprod", "crossprod", "star", or "chunked"}

\item{chunk_size}{numeric. If greater than 1 is Number of rows per chunk when \code{transpose_method = "chunked"}. Defaults to 1000. If less than or equal to 1, the entire matrix is processed in a single chunk.}

\item{verbose}{logical.  If TRUE, print progress through stages of algorithm}

\item{force_symmetric}{logical. If TRUE, forces the output matrix to be symmetric using Matrix::forceSymmetric.  This can be helpful when using chunked multiplication to ensure the final result is exactly symmetric despite potential numerical issues.  Defaults to TRUE.}
}
\description{
Compute the transpose multiplication for the relatedness matrix
}
\details{
The algorithms and methodologies used in this function are further discussed and exemplified in the vignette titled "examplePedigreeFunctions". For more advanced scenarios and detailed explanations, consult this vignette.
}
