% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildComponent.R
\name{.loadOrComputeInverseDiagonal}
\alias{.loadOrComputeInverseDiagonal}
\title{Load or compute the inverse diagonal matrix}
\usage{
.loadOrComputeInverseDiagonal(r, isChild, checkpoint_files, config)
}
\arguments{
\item{r}{The relatedness matrix.}

\item{config}{A list containing configuration parameters such as `resume`, `verbose`, and `saveable`.}
}
\value{
The computed inverse diagonal matrix.
}
\description{
Load or compute the inverse diagonal matrix
}
\keyword{internal}
