% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildComponent.R
\name{.loadOrComputeIsPar}
\alias{.loadOrComputeIsPar}
\title{Load or compute the isPar matrix}
\usage{
.loadOrComputeIsPar(iss, jss, parVal, ped, checkpoint_files, config)
}
\arguments{
\item{iss}{The row indices of the sparse matrix.}

\item{jss}{The column indices of the sparse matrix.}

\item{parVal}{The value to assign to the non-zero elements of the sparse matrix.}

\item{ped}{The pedigree dataset.}

\item{checkpoint_files}{A list of checkpoint file paths.}

\item{config}{A list containing configuration parameters such as `resume`, `verbose`, and `saveable`.}
}
\description{
Load or compute the isPar matrix
}
\keyword{internal}
