% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildComponent.R
\name{.loadOrComputeParList}
\alias{.loadOrComputeParList}
\title{parent-child adjacency data}
\usage{
.loadOrComputeParList(
  checkpoint_files,
  config,
  nr,
  ped = NULL,
  parList = NULL,
  lens = NULL
)
}
\arguments{
\item{checkpoint_files}{A list of checkpoint file paths.}

\item{config}{A list containing configuration parameters such as `resume`, `verbose`, and `saveable`.}

\item{nr}{The number of rows in the sparse matrix.}

\item{ped}{a pedigree dataset.  Needs ID, momID, and dadID columns}

\item{parList}{A list of parent-child adjacency data.}

\item{lens}{A vector of lengths for each parent-child relationship.}
}
\value{
A list containing the parent-child adjacency data either loaded from a checkpoint or initialized.
}
\description{
parent-child adjacency data
}
\keyword{internal}
