% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpBuildComponent.R
\name{.subsetKeepIds}
\alias{.subsetKeepIds}
\title{Subset output to requested IDs}
\usage{
.subsetKeepIds(
  component,
  keep_ids = NULL,
  available_ids,
  config,
  verbose_message = "Subsetting to \%d target individuals\\n",
  drop = FALSE
)
}
\arguments{
\item{component}{A component to subset.}

\item{keep_ids}{Character vector of IDs to retain.}

\item{available_ids}{Character vector of IDs available in \code{x}.}

\item{verbose_message}{Character. Message prefix to print when \code{config$verbose == TRUE}.}

\item{drop}{logical. Passed to \code{[} when subsetting matrices.}
}
\description{
Subset output to requested IDs
}
\keyword{internal}
