% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evenInsert.R
\name{evenInsert}
\alias{evenInsert}
\title{evenInsert
A function to insert m elements evenly into a length n vector.}
\usage{
evenInsert(m, n)
}
\arguments{
\item{m}{A numeric vector of length less than or equal to n. The elements to be inserted.}

\item{n}{A numeric vector. The vector into which the elements of m will be inserted.}
}
\value{
Returns a numeric vector with the elements of m evenly inserted into n.
}
\description{
evenInsert
A function to insert m elements evenly into a length n vector.
}
\details{
The function takes two vectors, m and n, and inserts the elements of m evenly into n.
If the length of m is greater than the length of n, the vectors are swapped, and the insertion proceeds.
The resulting vector is a combination of m and n, with the elements of m evenly distributed within n.
}
\seealso{
\code{\link{SimPed}} for the main function that uses this supporting function.
}
