% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPedigree.R
\name{extract_info}
\alias{extract_info}
\title{Extract Information from Line}
\usage{
extract_info(line, type)
}
\arguments{
\item{line}{A character string representing a line from a GEDCOM file.}

\item{type}{A character string representing the type of information to extract.}
}
\value{
A character string with the extracted information.
}
\description{
This function extracts information from a line based on a specified type.
}
\keyword{internal}
