% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famSizeCal.R
\name{famSizeCal}
\alias{famSizeCal}
\title{famSizeCal
A function to calculate the total number of individuals in a pedigree given parameters. This is a supporting function for function \code{simulatePedigree}}
\usage{
famSizeCal(kpc, Ngen, marR)
}
\arguments{
\item{kpc}{Number of kids per couple (integer >= 2).}

\item{Ngen}{Number of generations (integer >= 1).}

\item{marR}{Mating rate (numeric value ranging from 0 to 1).}
}
\value{
Returns a numeric value indicating the total pedigree size.
}
\description{
famSizeCal
A function to calculate the total number of individuals in a pedigree given parameters. This is a supporting function for function \code{simulatePedigree}
}
