% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePedigree.R
\name{findBiggest}
\alias{findBiggest}
\alias{findbiggest}
\title{This function finds the biggest families in a pedigree. It is supposed to be
used internally by the \code{summarize_pedigree} function.}
\usage{
findBiggest(foo_summary_dt, nbiggest, n_foo)
}
\arguments{
\item{foo_summary_dt}{A data.table containing the summary statistics.}

\item{nbiggest}{Integer. Number of largest lineages to return (sorted by count).}

\item{n_foo}{An integer specifying the number of individuals in the summary.}
}
\value{
a data.table containing the biggest families in the pedigree.
}
\description{
This function finds the biggest families in a pedigree. It is supposed to be
used internally by the \code{summarize_pedigree} function.
}
