% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePedigree.R
\name{findBiggest}
\alias{findBiggest}
\alias{findbiggest}
\title{Function to find the biggest families in a pedigree
This function finds the biggest families in a pedigree. It is supposed to be
used internally by the \code{summarize_pedigree} function.}
\usage{
findBiggest(foo_summary_dt, n_fooest = 5, n_foo_total = nrow(foo_summary_dt))
}
\arguments{
\item{foo_summary_dt}{A data.table containing the summary statistics.}

\item{n_fooest}{An integer specifying the number of individuals in the summary.}

\item{n_foo_total}{An integer specifying the total number of individuals in the summary.}
}
\value{
a data.table containing the biggest families in the pedigree.
}
\description{
Function to find the biggest families in a pedigree
This function finds the biggest families in a pedigree. It is supposed to be
used internally by the \code{summarize_pedigree} function.
}
