% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePedigree.R
\name{findOldest}
\alias{findOldest}
\alias{findoldest}
\title{This function finds the oldest families in a pedigree. It is supposed to be used internally by the \code{summarize_pedigree} function.}
\usage{
findOldest(foo_summary_dt, byr, n_oldest, n_foo)
}
\arguments{
\item{foo_summary_dt}{A data.table containing the summary statistics.}

\item{byr}{Character. Optional column name for birth year. Used to determine the oldest lineages.}

\item{n_oldest}{Integer. Number of oldest lineages to return (sorted by birth year).}

\item{n_foo}{An integer specifying the number of individuals in the summary.}
}
\value{
a data.table containing the oldest families in the pedigree.
}
\description{
This function finds the oldest families in a pedigree. It is supposed to be used internally by the \code{summarize_pedigree} function.
}
