% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{fitComponentModel}
\alias{fitComponentModel}
\title{Fit the estimated variance components of a model to covariance data}
\usage{
fitComponentModel(covmat, ...)
}
\arguments{
\item{covmat}{the covariance matrix of the raw data, possibly blockwise.}

\item{...}{Comma-separated relatedness component matrices.}
}
\description{
Fit the estimated variance components of a model to covariance data
}
\details{
Returns a regression (linear model fitted with \code{lm}).
The coeficients of the regression are the estimated variance components.
}
