% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmxPedigrees.R
\name{fitPedigreeModel}
\alias{fitPedigreeModel}
\title{fitPedigreeModel
Fit an OpenMx pedigree model to observed data.
This function constructs and fits an OpenMx model for a pedigree using specified variance components and family group models.}
\usage{
fitPedigreeModel(
  model_name = "PedigreeModel",
  vars = c(ad2 = 0.5, dd2 = 0.3, cn2 = 0.2, ce2 = 0.4, mt2 = 0.1, am2 = 0.25, ee2 = 0.6),
  data,
  group_models = NULL
)
}
\description{
fitPedigreeModel
Fit an OpenMx pedigree model to observed data.
This function constructs and fits an OpenMx model for a pedigree using specified variance components and family group models.
}
