% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{hazard}
\alias{hazard}
\title{Simulated pedigree with two extended families and an age-related hazard}
\format{
A data frame with 43 rows and 14 variables
}
\usage{
data(hazard)
}
\description{
A dataset simulated to have an age-related hazard.
There are two extended families that are sampled from the same population.
}
\details{
The variables are as follows:

\itemize{
  \item \code{FamID}:  ID of the extended family
  \item \code{ID}:  Person identification variable
  \item \code{sex}:  Sex of the ID: 1 is female; 0 is male
  \item \code{dadID}:  ID of the father
  \item \code{momID}:  ID of the mother
  \item \code{affected}:  logical.  Whether the person is affected or not
  \item \code{DA1}:  Binary variable signifying the meaninglessness of life
  \item \code{DA2}:  Binary variable signifying the fundamental unknowability of existence
  \item \code{birthYr}:  Birth year for person
  \item \code{onsetYr}:  Year of onset for person
  \item \code{deathYr}:  Death year for person
  \item \code{available}:  logical.  Whether
  \item \code{Gen}:  Generation of the person
  \item \code{proband}:  logical.  Whether the person is a proband or not
}
}
\keyword{datasets}
