% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildComponent.R
\name{initializeCheckpoint}
\alias{initializeCheckpoint}
\title{Initialize checkpoint files}
\usage{
initializeCheckpoint(
  config = list(verbose = FALSE, saveable = FALSE, resume = FALSE, save_path =
    "checkpoint/")
)
}
\description{
Initialize checkpoint files
}
\keyword{internal}
