% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpTwins.R
\name{isTwin}
\alias{isTwin}
\title{Determine isTwin Status}
\usage{
isTwin(ped)
}
\arguments{
\item{ped}{pedigree data frame}
}
\value{
A logical vector indicating, for each row of \code{ped}, whether
  \code{twinID} is non-\code{NA}.
}
\description{
Determine isTwin Status
}
\keyword{internal}
