% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildComponent.R
\name{loadOrComputeCheckpoint}
\alias{loadOrComputeCheckpoint}
\title{Load or compute a checkpoint}
\usage{
loadOrComputeCheckpoint(
  file,
  compute_fn,
  config,
  message_resume = NULL,
  message_compute = NULL
)
}
\arguments{
\item{file}{The file path to load the checkpoint from.}

\item{compute_fn}{The function to compute the checkpoint if it doesn't exist.}

\item{config}{A list containing configuration parameters such as `resume`, `verbose`, and `saveable`.}

\item{message_resume}{Optional message to display when resuming from a checkpoint.}

\item{message_compute}{Optional message to display when computing the checkpoint.}
}
\value{
The loaded or computed checkpoint.
}
\description{
Load or compute a checkpoint
}
\keyword{internal}
