% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildComponent.R
\name{loadOrComputeCheckpoint}
\alias{loadOrComputeCheckpoint}
\title{Load or compute a checkpoint}
\usage{
loadOrComputeCheckpoint(
  file,
  compute_fn,
  config,
  message_resume = NULL,
  message_compute = NULL,
  compress = TRUE
)
}
\arguments{
\item{file}{The file path to load the checkpoint from.}

\item{compute_fn}{The function to compute the checkpoint if it doesn't exist.}

\item{config}{A list containing configuration parameters such as `resume`, `verbose`, and `saveable`.}

\item{message_resume}{Optional message to display when resuming from a checkpoint.}

\item{message_compute}{Optional message to display when computing the checkpoint.}

\item{compress}{a logical specifying whether saving to a named file is to use "gzip" compression, or one of "gzip", "bzip2", "xz" or "zstd" to indicate the type of compression to be used. Ignored if file is a connection.}
}
\value{
The loaded or computed checkpoint.
}
\description{
Load or compute a checkpoint
}
\keyword{internal}
