% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpbuildmxPedigrees.R
\name{make_clean_personids}
\alias{make_clean_personids}
\title{Make Clean IDs for OpenMx}
\usage{
make_clean_personids(ids)
}
\arguments{
\item{ids}{A vector of IDs to be cleaned.}
}
\value{
A vector of cleaned IDs suitable for use in OpenMx models.
}
\description{
This function takes a vector of IDs and returns a cleaned version suitable for use in OpenMx models. It replaces any illegal characters (such as '.') with underscores and ensures that the IDs are valid R variable names. This is important because OpenMx does not allow certain characters in matrix or variable names, which can lead to errors when building models.
}
