% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPedigree.R
\name{ped2add}
\alias{ped2add}
\title{Take a pedigree and turn it into an additive genetics relatedness matrix}
\usage{
ped2add(ped, max.gen = Inf, sparse = FALSE, verbose = FALSE, gc = FALSE)
}
\arguments{
\item{ped}{a pedigree data file.  Needs ID, momID, and dadID columns}

\item{max.gen}{the maximum number of generations to compute
(e.g., only up to 4th degree relatives).  The default of Inf uses as many
generations as there are in the data.}

\item{sparse}{logical.  If TRUE, use and return sparse matrices from Matrix package}

\item{verbose}{logical  If TRUE, print progress through stages of algorithm}

\item{gc}{logical. If TRUE, do frequent garbage collection via \code{\link{gc}} to save memory}
}
\description{
Take a pedigree and turn it into an additive genetics relatedness matrix
}
\details{
source examplePedigreeFunctions
}
