% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPedigree.R
\name{ped2com}
\alias{ped2com}
\title{Take a pedigree and turn it into a relatedness matrix}
\usage{
ped2com(
  ped,
  component,
  max.gen = Inf,
  sparse = FALSE,
  verbose = FALSE,
  gc = FALSE,
  flatten_diag = FALSE
)
}
\arguments{
\item{ped}{a pedigree dataset.  Needs ID, momID, and dadID columns}

\item{component}{character.  Which component of the pedigree to return.  See Details.}

\item{max.gen}{the maximum number of generations to compute
(e.g., only up to 4th degree relatives).  The default of Inf uses as many
generations as there are in the data.}

\item{sparse}{logical.  If TRUE, use and return sparse matrices from Matrix package}

\item{verbose}{logical  If TRUE, print progress through stages of algorithm}

\item{gc}{logical. If TRUE, do frequent garbage collection via \code{\link{gc}} to save memory}

\item{flatten_diag}{Logical. The default is FALSE. If TRUE, overwrites the diagonal of the final relatedness matrix with ones.}
}
\description{
Take a pedigree and turn it into a relatedness matrix
}
\details{
source examplePedigreeFunctions
}
