% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{ped2graph}
\alias{ped2graph}
\title{Turn a pedigree into a graph}
\usage{
ped2graph(
  ped,
  personID = "ID",
  momID = "momID",
  dadID = "dadID",
  famID = "famID"
)
}
\arguments{
\item{ped}{a pedigree dataset.  Needs ID, momID, and dadID columns}

\item{personID}{character.  Name of the column in ped for the person ID variable}

\item{momID}{character.  Name of the column in ped for the mother ID variable}

\item{dadID}{character.  Name of the column in ped for the father ID variable}

\item{famID}{character.  Name of the column to be created in ped for the family ID variable}
}
\value{
A graph
}
\description{
Turn a pedigree into a graph
}
\details{
The general idea of this function is to represent a pedigree as a graph using the igraph package.

Once in graph form, several common pedigree tasks become much simpler.
}
