% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGedcom.R
\name{postProcessGedcom}
\alias{postProcessGedcom}
\title{Post-process GEDCOM Data Frame}
\usage{
postProcessGedcom(
  df_temp,
  remove_empty_cols = TRUE,
  combine_cols = TRUE,
  add_parents = TRUE,
  skinny = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df_temp}{A data frame containing information about individuals.}

\item{remove_empty_cols}{A logical value indicating whether to remove columns with all missing values.}

\item{combine_cols}{A logical value indicating whether to combine columns with duplicate values.}

\item{add_parents}{A logical value indicating whether to add parents to the data frame.}

\item{skinny}{A logical value indicating whether to return a skinny data frame.}

\item{verbose}{A logical value indicating whether to print messages.}
}
\value{
A data frame with processed information.
}
\description{
Post-process GEDCOM Data Frame
}
