% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{potter}
\alias{potter}
\title{Fictional pedigree data on a wizarding family}
\format{
A data frame (and ped object) with 36 rows and 8 variables
}
\usage{
data(potter)
}
\description{
A dataset created purely from imagination that includes a subset of the Potter extended family.
}
\details{
The variables are as follows:

\itemize{
  \item \code{personID}:  Person identification variable
  \item \code{famID}: Family identification variable
  \item \code{name}:  Name of the person
  \item \code{gen}: Generation of the person
  \item \code{momID}:  ID of the mother
  \item \code{dadID}:  ID of the father
  \item \code{spouseID}: ID of the spouse
  \item \code{sex}:  Sex of the ID: 1 is male; 0 is female

}

IDs in the 100s \code{momID}s and \code{dadID}s are for people not in the dataset.
}
\keyword{datasets}
