% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{potter}
\alias{potter}
\title{Fictional pedigree data on a wizarding family}
\format{
A data frame (and ped object) with 24 rows and 5 variables
}
\usage{
data(potter)
}
\description{
A dataset created purely from imagination that includes a subset of the Potter extended family.
}
\details{
The variables are as follows:

\itemize{
  \item ID.  Person identification variable
  \item sex.  Sex of the ID: 1 is male; 0 is female
  \item maID.  ID of the mother
  \item paID.  ID of the father
  \item Name.  Name of the person
}

Negative \code{maID}s and \code{paID}s are for people not in the dataset.
}
\keyword{datasets}
