% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGedcom.R
\name{processParents}
\alias{processParents}
\title{Process parents information}
\usage{
processParents(df_temp, datasource)
}
\arguments{
\item{df_temp}{A data frame containing information about individuals.}
}
\value{
A data frame with added momID and dadID columns.
}
\description{
This function processes the dataframe to add momID and dadID columns.
}
\keyword{internal}
