% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPedigree.R
\name{readGedcom}
\alias{readGedcom}
\title{Read a GEDCOM File}
\usage{
readGedcom(
  file_path,
  verbose = FALSE,
  add_parents = TRUE,
  remove_empty_cols = TRUE
)
}
\arguments{
\item{file_path}{The path to the GEDCOM file.}

\item{verbose}{A logical value indicating whether to print messages.}

\item{add_parents}{A logical value indicating whether to add parents to the data frame.}

\item{remove_empty_cols}{A logical value indicating whether to remove columns with all missing values.}
}
\value{
A data frame containing information about individuals, with the following columns:
- `id`: ID of the individual
- `firstname`: First name of the individual
- `lastname`: Last name of the individual
- `birthdate`: Birth date of the individual
- `birthplace`: Birthplace of the individual
- `birthlat`: Latitude of the birthplace
- `birthlong`: Longitude of the birthplace
- `deathdate`: Death date of the individual
- `deathplace`: Place of death of the individual
- `deathlat`: Latitude of the place of death
- `deathlong`: Longitude of the place of death
- `sex`: Sex of the individual
- `FAMC`: ID(s) of the family where the individual is a child
- `FAMS`: ID(s) of the family where the individual is a spouse
- `momID`: ID of the individual's mother
- `dadID`: ID of the individual's father
- `MARNM`: Married Name
- `NPFX`: Name Prefix
- `CAUS`: Cause of death
- `title`: Title of the individual
- `occupation`: Occupation of the individual
- `religion`: Religion of the individual
- `education`: Education of the individual
}
\description{
This function reads a GEDCOM file and parses it into a structured data frame of individuals.
inpired by https://raw.githubusercontent.com/jjfitz/readgedcom/master/R/read_gedcom.R
}
