% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPedigree.R
\name{readGedcom}
\alias{readGedcom}
\title{Read a GEDCOM File}
\usage{
readGedcom(
  file_path,
  verbose = FALSE,
  add_parents = TRUE,
  remove_empty_cols = TRUE
)
}
\arguments{
\item{file_path}{The path to the GEDCOM file.}

\item{verbose}{A logical value indicating whether to print messages.}

\item{add_parents}{A logical value indicating whether to add parents to the data frame.}

\item{remove_empty_cols}{A logical value indicating whether to remove columns with all missing values.}
}
\value{
A data frame containing information about individuals, with the following columns:
\itemize{
  \item{id}{ID of the individual.}
  \item{firstname}{First name of the individual.}
  \item{lastname}{Last name of the individual.}
  \item{birthdate}{Birth date of the individual.}
  \item{birthplace}{Birthplace of the individual.}
  \item{birthlat}{Latitude of the birthplace.}
  \item{birthlong}{Longitude of the birthplace.}
  \item{deathdate}{Death date of the individual.}
  \item{deathplace}{Place of death of the individual.}
  \item{deathlat}{Latitude of the place of death.}
  \item{deathlong}{Longitude of the place of death.}
  \item{sex}{Sex of the individual.}
  \item{FAMC}{ID(s) of the family where the individual is a child.}
  \item{FAMS}{ID(s) of the family where the individual is a spouse.}
  \item{momID}{ID of the individual's mother.}
  \item{dadID}{ID of the individual's father.}
}
}
\description{
This function reads a GEDCOM file and parses it into a structured data frame of individuals.
adapted from: https://raw.githubusercontent.com/jjfitz/readgedcom/master/R/read_gedcom.R
}
