% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPedigree.R
\name{readWikifamilytree}
\alias{readWikifamilytree}
\title{Read Wiki Family Tree}
\usage{
readWikifamilytree(text)
}
\arguments{
\item{text}{A character string containing the text of a family tree in wiki format.}
}
\description{
Read Wiki Family Tree
}
