% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{related_coef}
\alias{related_coef}
\title{Relatedness Coefficient Calculation}
\usage{
related_coef(generations = 2, path = NULL, both = TRUE, ...)
}
\arguments{
\item{generations}{A number}

\item{path}{A number}

\item{both}{A number}
}
\value{
Relatedness Coefficient \code{coef}
}
\description{
Relatedness Coefficient Calculation based on Wright (1922)
}
\details{
\eqn{r_{bc} = \sum \left(\frac{1}{2}\right)^{n+n'+1} (1+f_a)}
where the relatedness coefficient between two people (b & c) is defined in relation to their common ancestors.
}
\examples{


}
