% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpGeneric.R
\name{relatedness}
\alias{relatedness}
\title{relatedness (Deprecated)}
\usage{
relatedness(...)
}
\arguments{
\item{...}{Arguments to be passed to `inferRelatedness`.}
}
\value{
The same result as calling `inferRelatedness`.
}
\description{
When calling this function, a warning will be issued about its deprecation.
}
\details{
This function is a wrapper around the new `inferRelatedness` function.
`relatedness` has been deprecated, and it's advised to use `inferRelatedness` directly.
}
\examples{
\dontrun{
# This is an example of the deprecated function:
relatedness(...)
# It is recommended to use:
inferRelatedness(...)
}
}
\seealso{
\code{\link{inferRelatedness}} for the updated function.
}
\keyword{deprecated}
