% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkParents.R
\name{repairParentIDs}
\alias{repairParentIDs}
\title{Repair Parent IDs}
\usage{
repairParentIDs(ped, verbose = FALSE)
}
\arguments{
\item{ped}{A dataframe representing the pedigree data with columns 'ID', 'dadID', and 'momID'.}

\item{verbose}{A logical flag indicating whether to print progress and validation messages to the console.}
}
\value{
A corrected pedigree
}
\description{
This function repairs parent IDs in a pedigree.
}
