% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatePedigree.R
\name{repairPedigree}
\alias{repairPedigree}
\title{Repair Pedigree}
\usage{
repairPedigree(ped, repair_funs = NULL, ...)
}
\arguments{
\item{ped}{A pedigree object.}

\item{repair_funs}{A list of functions to repair the pedigree.}
}
\value{
A corrected pedigree.
}
\description{
This function applies a list of repair functions sequentially to a pedigree.
}
