% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSex.R
\name{repairSex}
\alias{repairSex}
\title{Repairs Sex Coding in a Pedigree Dataframe}
\usage{
repairSex(ped, verbose = FALSE, code_male = NULL, code_female = NULL)
}
\arguments{
\item{ped}{A dataframe representing the pedigree data with a 'sex' column.}

\item{verbose}{A logical flag indicating whether to print progress and validation messages to the console.}

\item{code_male}{The code used to represent males in the 'sex' column. If both male and female are NULL, no recoding is performed.}

\item{code_female}{The code used to represent females in the 'sex' column. If both male and female are NULL, no recoding is performed.}
}
\value{
A dataframe where the sex coding has been repaired.
}
\description{
This function serves as a wrapper around `checkSex` to specifically handle
the repair of the sex coding in a pedigree dataframe.
}
\examples{
\dontrun{
ped <- data.frame(ID = c(1, 2, 3), sex = c("M", "F", "M"))
repairSex(ped, code_male = "M", verbose = TRUE)
}
}
\seealso{
\code{\link{checkSex}}
}
