% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{royal92}
\alias{royal92}
\title{Royal pedigree data from 1992}
\format{
A data frame with 3110 observations
}
\usage{
data(royal92)
}
\description{
A dataset created by Denis Reid from the Royal Families of Europe.
}
\details{
The variables are as follows:
id,momID,dadID,name,sex,birth_date,death_date,attribute_title
\itemize{
  \item \code{id}:  Person identification variable
  \item \code{momID}:  ID of the mother
  \item \code{dadID}:  ID of the father
  \item \code{name}:  Name of the person
  \item \code{sex}: Biological sex
  \item \code{birth_date}:  Date of birth
  \item \code{death_date}:  Date of death
  \item \code{attribute_title}:  Title of the person

}
}
\keyword{datasets}
