% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famSizeCal.R
\name{sizeAllGens}
\alias{sizeAllGens}
\title{sizeAllGens
An internal supporting function for function \code{allGens}}
\usage{
sizeAllGens(kpc, Ngen, marR)
}
\arguments{
\item{kpc}{kids per couple}

\item{Ngen}{number of generations}

\item{marR}{Mating rate.}
}
\value{
Returns a vector including the number of individuals in every generation.
}
\description{
sizeAllGens
An internal supporting function for function \code{allGens}
}
