% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyModel.R
\name{vech}
\alias{vech}
\title{vech
Create the half-vectorization of a matrix}
\usage{
vech(x)
}
\arguments{
\item{x}{a matrix, the half-vectorization of which is desired}
}
\value{
A vector containing the lower triangle of the matrix, including the diagonal.
}
\description{
vech
Create the half-vectorization of a matrix
}
\details{
This function returns the vectorized form of the lower triangle of a matrix, including the diagonal.
The upper triangle is ignored with no checking that the provided matrix is symmetric.
}
\examples{

vech(matrix(c(1, 0.5, 0.5, 1), nrow = 2, ncol = 2))

}
